refactor: Conference 관련 모델 mapper 분리
