+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/conference/ConferenceRepositoryImpl.kt
@@ -1,11 +1,14 @@
 package com.emmsale.data.conference
+
 import com.emmsale.data.common.ApiResult
 import com.emmsale.data.common.handleApi
 import com.emmsale.data.conference.dto.ConferenceApiModel
-import com.emmsale.data.conference.dto.toData
+import com.emmsale.data.conference.mapper.toApiModel
+import com.emmsale.data.conference.mapper.toData
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
+
 class ConferenceRepositoryImpl(
     private val dispatcher: CoroutineDispatcher = Dispatchers.IO,
     private val conferenceService: ConferenceService,
@@ -20,11 +23,11 @@ class ConferenceRepositoryImpl(
         handleApi(
             execute = {
                 conferenceService.getEvents(
-                    category.text,
-                    startDate,
-                    endDate,
-                    statuses.toTexts(),
-                    tags,
+                    category = category.text,
+                    startDate = startDate,
+                    endDate = endDate,
+                    statuses = statuses.toApiModel(),
+                    tags = tags,
                 )
             },
             mapToDomain = List<ConferenceApiModel>::toData,
