+++ android/2023-emmsale/app/src/main/java/com/emmsale/data/conference/ConferenceStatus.kt
@@ -1,15 +1,7 @@
 package com.emmsale.data.conference
 
-enum class ConferenceStatus(val text: String) {
-    IN_PROGRESS("IN_PROGRESS"),
-    SCHEDULED("UPCOMING"),
-    ENDED("ENDED"),
-    ;
-
-    companion object {
-        fun from(status: String): ConferenceStatus = values().find { it.text == status }
-            ?: throw IllegalArgumentException("${status}는 올바르지 않은 행사 상태입니다.")
-    }
+enum class ConferenceStatus {
+    IN_PROGRESS,
+    SCHEDULED,
+    ENDED,
 }
-
-fun List<ConferenceStatus>.toTexts(): List<String> = map { it.text }
