+++ android/2023-emmsale/app/src/main/res/values/strings.xml
@@ -77,8 +77,8 @@
     <string name="eventfilter_conference_duration_title">기간</string>
     <string name="eventfilter_apply">적용하기</string>
     <string name="eventfilter_conference_duration_divider_mark">~</string>
-    <string name="eventfilter_duration_date_format">%02d.%02d</string>
-    <string name="eventfilter_duration_date_last_format">&#160;- %02d.%02d</string>
+    <string name="eventfilter_duration_date_format">%02d.%02d.%02d</string>
+    <string name="eventfilter_duration_date_last_format">&#160;- %02d.%02d.%02d</string>
     <string name="eventfilter_clear">초기화</string>
     <string name="eventfilter_clear_title">필터 초기화</string>
     <string name="eventfilter_clear_message">선택된 필터를 모두 초기화 할까요?</string>
